+++ packages/nft-checkout/src/const/tokens.ts
@@ -11,7 +11,7 @@ export const PANCAKE_SWAP_TESTNET_TOKEN_LIST: Token[] = [
     'Ethereum Token',
     'ETH',
     18,
-    'https://raw.githubusercontent.com/MetaMask/metamask-extension/develop/app/images/eth_logo.svg',
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/eth-logo.png',
   ),
   newToken(
     chapelChain,
@@ -19,7 +19,7 @@ export const PANCAKE_SWAP_TESTNET_TOKEN_LIST: Token[] = [
     'Wrapped Binance Testnet Token',
     'WBNB',
     18,
-    'https://raw.githubusercontent.com/MetaMask/metamask-extension/develop/app/images/bnb.png',
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/bnb-logo.png',
   ),
   newToken(
     chapelChain,
@@ -27,7 +27,7 @@ export const PANCAKE_SWAP_TESTNET_TOKEN_LIST: Token[] = [
     'USDCoin',
     'USDC',
     18,
-    'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png',
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/usdc-logo.png',
   ),
   newToken(
     chapelChain,
@@ -35,7 +35,7 @@ export const PANCAKE_SWAP_TESTNET_TOKEN_LIST: Token[] = [
     'Tether USD',
     'USDT',
     18,
-    'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0xdAC17F958D2ee523a2206206994597C13D831ec7/logo.png',
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/usdt-logo.png',
   ),
 ]
 
@@ -46,7 +46,7 @@ export const TRADER_JOE_SWAP_TESTNET_TOKEN_LIST: Token[] = [
     'Wrapped AVAX',
     'WAVAX',
     18,
-    'https://raw.githubusercontent.com/traderjoe-xyz/joe-tokenlists/main/logos/0xd00ae08403B9bbb9124bB305C09058E32C39A48c/logo.png',
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/wavax-logo.png',
   ),
   newToken(
     fujiChain,
@@ -54,6 +54,6 @@ export const TRADER_JOE_SWAP_TESTNET_TOKEN_LIST: Token[] = [
     'USD Coin',
     'USDC',
     6,
-    'https://raw.githubusercontent.com/traderjoe-xyz/joe-tokenlists/main/logos/0xB6076C93701D6a07266c31066B298AeC6dd65c2d/logo.png',
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/usdc-logo.png',
   ),
 ]