+++ packages/nft-checkout/src/enums/operation-event-bus.ts
@@ -1,3 +1,4 @@
 export enum OperationEventBusEvents {
   Initiated = 'initiated',
+  StatusChanged = 'status-changed',
 }