+++ packages/nft-checkout/src/types/operation-event-bus.ts
@@ -3,22 +3,22 @@ import type { BridgeChain } from '@rarimo/shared'
 import { OperationEventBusEvents } from '@/enums'
 
 import type { Target } from './operation'
+import type { CheckoutOperationStatus } from './operation'
 
-export type OperationInitiatedEventPayload = {
+export type OperationEventPayload = {
   isInitiated: boolean
-  chainFrom: BridgeChain
+  chainFrom?: BridgeChain
   target?: Target
+  status: CheckoutOperationStatus
 }
 
 export type OperationEventMap = {
-  [OperationEventBusEvents.Initiated]: {
-    isInitiated: boolean
-    chainFrom: BridgeChain
-    target?: Target
-  }
+  [OperationEventBusEvents.Initiated]: OperationEventPayload
+  [OperationEventBusEvents.StatusChanged]: OperationEventPayload
 }
 
 export interface OperationSubscriber {
-  onInitiated(cb: (e?: OperationInitiatedEventPayload) => void): void
+  onInitiated(cb: (e?: OperationEventPayload) => void): void
+  onStatusChanged(cb: (e?: OperationEventPayload) => void): void
   clearHandlers(): void
 }