+++ packages/react-nft-checkout/src/hooks/useCheckoutOperation.ts
@@ -32,16 +32,27 @@ export const useCheckoutOperation = ({
       return {
         isInitiated: checkoutOperation?.isInitialized,
         chainFrom: checkoutOperation?.chainFrom,
+        status: checkoutOperation?.status,
       }
     })
 
   const setListeners = useCallback(() => {
     if (!checkoutOperation) return
 
-    checkoutOperation.onInitiated(e => {
+    checkoutOperation.onInitiated(payload => {
+      const { isInitiated, chainFrom } = payload ?? {}
+
       setCheckoutOperationReactiveState(prev => ({
         ...prev,
-        ...(e || {}),
+        chainFrom,
+        isInitiated,
+      }))
+    })
+
+    checkoutOperation.onStatusChanged(payload => {
+      setCheckoutOperationReactiveState(prev => ({
+        ...prev,
+        status: payload?.status,
       }))
     })
   }, [checkoutOperation])